% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{plot.transformer}
\alias{plot.transformer}
\title{Function plotting the transformed data.}
\usage{
\method{plot}{transformer}(object, new_data = NULL, plot_all = FALSE, ...)
}
\arguments{
\item{object}{transformer object.}

\item{new_data}{new data having the same features to the data creating the transformer object. If new_data is NULL, the fitted data of the object is plot. If new_data is a dataset, this dataset will be transformed by the object and plotted.}

\item{plot_all}{binary flag to plot all fitted data and transformed new_data if new_data is not NULL. If 'plot_all' is TRUE and the 'color' is set, 'color' should be set to the combination length accordingly.}

\item{color}{optional colors for each data point, default is NULL, all data points are plotted in black. If labels is set, the data points' colors are encoded accordingly.}
}
\description{
This function plots the transformed data.
}
\details{
This function
}
