% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{plottrans.transformer}
\alias{plottrans.transformer}
\title{Function transforming and plotting the transformed data.}
\usage{
\method{plottrans}{transformer}(object, new_data, point_label = NULL, ...)
}
\arguments{
\item{object}{transformer object.}

\item{new_data}{new data having the same features to the data instantiate the transformer object.}

\item{point_label}{label for each data point, default is NULL, all data points are plotted in black. If labels is setted, the data points' color is encoded accordingly.}
}
\description{
This function transforms data and plots the transformed data.
}
\details{
This function
}
