% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{print.transformer}
\alias{print.transformer}
\title{Function printing the output of the transformer object}
\usage{
\method{print}{transformer}(object, ...)
}
\description{
Function printing the output of the transformer object
}
