% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{summary.transformer}
\alias{summary.transformer}
\title{Function printing the summary of intermediate algorithm output of the transformer object}
\usage{
\method{summary}{transformer}(object, ...)
}
\description{
Function printing the summary of intermediate algorithm output of the transformer object
}
