% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transform.transformer}
\alias{transform.transformer}
\title{Function transforming new data by the existing transformer object.}
\usage{
\method{transform}{transformer}(object, newdata)
}
\arguments{
\item{object}{transformer object.}

\item{new_data}{new data having the same features to the data instantiate the transformer object.}
}
\value{
a dataset of transformed data.
}
\description{
This function return transformed data for new data.
}
\details{
This applies the transformer object metadata to transform new data.
}
\examples{
data(iris)
data <- iris[sample(1:nrow(iris)),]
idx_train <- 1:140
idx_test <- 141:nrow(data)
x_trans <- transformer.kpca(iris[idx_train,1:4])
transform(x_trans, iris[idx_test,1:4])

}
