% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{transformer.kpca}
\alias{transformer.kpca}
\title{Function creating a transformer object by Kernel PCA algorithm.}
\usage{
transformer.kpca(
  x,
  components = 2,
  center = FALSE,
  scaling = FALSE,
  handle_discrete = FALSE,
  kernel = "rbfdot",
  sigma = 0.1
)
}
\arguments{
\item{components}{positive number of components for the transformed data.}

\item{center}{boolean value to scale data. Parameter is passed to base::scale.}

\item{scaling}{boolean value to scale data. Parameter is passed to base::scale.}

\item{handle_discrete}{boolean value to handle discrete features in data.}

\item{kernel}{string to indicate the kernel name. Currently, 'rbfdot' is allowed.}

\item{sigma}{numeric value to indicate the inverse kernel width for the Radial Basis kernel function "rbfdot".}
}
\value{
transformer.pca return a class "transformer" containing the following componenets:
}
\description{
This function create a transformer object fitted by data.
}
\details{
This calculation is created based on stat::prcomp function with some adjustments to fit into the purpose of the package. It creates a transformer object including several attribute to perform other functionality.
}
\examples{
data(iris)
x_trans <- transformer.kpca(iris[,1:4])

}
