% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{transformer.pca}
\alias{transformer.pca}
\title{Function creating a transformer object using PCA algorithm.}
\usage{
transformer.pca(
  x,
  components = 2,
  center = FALSE,
  scaling = FALSE,
  handle_category = NULL
)
}
\arguments{
\item{components}{positive number of components for the transformed data.}

\item{center}{logical value to scale data. Parameter is passed to base::scale.}

\item{scaling}{logical value to scale data. Parameter is passed to base::scale.}

\item{handle_category}{character value to handle categorical features. The accepted values are 'label', 'onehot', and 'ignore'. Default value is NULL, if dataset contains character fields, the function return error. .}
}
\value{
Return "transformer" class.
}
\description{
This function return a transformer object fitted by data.
}
\details{
The function is created based on stat::prcomp function with some adjustments to fit into the purpose of the package. It includes preprocessing data (scaling, categorical handling) before transforming data. It creates a transformer object with attributes to perform other functionalities.
}
\examples{
data(iris)
pca <- transformer.pca(iris[,1:4])
print(pca)
}
