% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{inverse}
\alias{inverse}
\alias{inverse.default}
\alias{inverse.transformer}
\title{Function inversing the tranformed data back to the original data.}
\usage{
inverse(object, ...)

\method{inverse}{default}(object, data, ...)

\method{inverse}{transformer}(object, data, ...)
}
\arguments{
\item{object}{transformer object.}

\item{data}{dataframe being inversed.}
}
\value{
a dataset after inversing.
}
\description{
This function inverses the transformed data back to the original data based on the transformer object's attributes.
}
\details{
This function inverses the transformed data back to the original if using all set components. The inverse can happen partly for specific component to analyse the effect of components.
}
\section{Methods (by class)}{
\itemize{
\item \code{inverse(default)}: Default for inverse function.

\item \code{inverse(transformer)}: Inverse function for transformer object.

}}
\examples{
data(iris)
iris[1:10,1:4]
x_trans <- transformer.kpca(iris[1:10,1:4])
inverse(x_trans, x_trans$x)

}
