% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{plot.transformer}
\alias{plot.transformer}
\title{Function plotting the transformed data.}
\usage{
\method{plot}{transformer}(object, point_label = NULL, ...)
}
\arguments{
\item{object}{transformer object.}

\item{point_label}{optional label for each data point, default is NULL, all data points are plotted in black. If labels is set, the data points' colors are encoded accordingly.}
}
\description{
This function plots the transformed data.
}
\details{
This function
}
