% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{summary.transformer}
\alias{summary.transformer}
\title{Summarizing transformer object}
\usage{
\method{summary}{transformer}(object)
}
\arguments{
\item{object}{transformer object.}
}
\value{
Return a list of transformer object's attributes.
}
\description{
summary method for class 'transformer'.
}
\details{
Object's attributes are printed.
}
