% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transform.transformer}
\alias{transform.transformer}
\title{Function transforming data by transformer object.}
\usage{
\method{transform}{transformer}(object, newdata)
}
\arguments{
\item{object}{transformer object.}

\item{new_data}{new data having the same features to the data instantiate the transformer object.}
}
\value{
a dataset of transformed data.
}
\description{
This function transforms new data by the existing object.
}
\details{
This applies the transformer object metadata to transform new data.
}
\examples{
data(iris)
data <- iris[sample(1:nrow(iris)),]
idx_train <- 1:140
idx_test <- 141:nrow(data)
pca <- transformer.pca(iris[idx_train,1:4])
test_transformed <- transform(pca, iris[idx_test,1:4])
head(test_transformed)

}
