% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{transformer.kpca}
\alias{transformer.kpca}
\title{Function creating a transformer object by PCA algorithm.}
\usage{
transformer.kpca(x, ...)
}
\arguments{
\item{component}{positive number of components for the transformed data.}

\item{center}{boolean value to scale data. Parameter is passed to base::scale.}

\item{scaling}{boolean value to scale data. Parameter is passed to base::scale.}

\item{handle_discrete}{boolean value to handle discrete features in data.}

\item{kernel}{string to indicate the kernel name, the allowed values are [].}

\item{sigma}{sigma...}

\item{th}{value to define warning threshold for eigenvalues.}
}
\value{
transformer.pca return a class "transformer" containing the following componenets:
}
\description{
This function return a transformer object fitted by data.
}
\details{
This calculation is created based on stat::prcomp function with some adjustments to fit into the purpose of the package. It creates a transformer object including several attribute to perform other functionality.
}
\examples{
data(iris)
x_trans <- transformer.kpca(iris[,1:4])

}
