% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{transformer.kpca}
\alias{transformer.kpca}
\title{Function creating a transformer object by Kernel PCA algorithm.}
\usage{
transformer.kpca(
  x,
  components = 2,
  center = FALSE,
  scaling = FALSE,
  handle_category = NULL,
  kernel = "rbfdot",
  sigma = 0.1
)
}
\arguments{
\item{components}{positive number of components for the transformed data.}

\item{center}{logical value to scale data. Parameter is passed to base::scale.}

\item{scaling}{logical value to scale data. Parameter is passed to base::scale.}

\item{handle_category}{character value to handle categorical features. The accepted values are 'label', 'onehot', and 'ignore'. Default value is NULL, if dataset contains character fields, the function return error. .}

\item{kernel}{character value to indicate the kernel name. Currently, 'rbfdot' is allowed.}

\item{sigma}{numeric value to indicate the inverse kernel width for the Radial Basis kernel function "rbfdot". The smaller of sigma, the more non-linear of the decision boundary.}
}
\value{
Return "transformer" class.
}
\description{
This function create a transformer object fitted by data.
}
\details{
The function is created based on skernlab::kpca function with some adjustments to fit into the purpose of the package. It includes preprocessing data (scaling, categorical handling) before transforming data. It creates a transformer object with attributes to perform other functionalities.
}
\examples{
data(iris)
kpca <- transformer.kpca(iris[,1:4])
print(kpca)
}
