% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{transformer.kpca}
\alias{transformer.kpca}
\title{Function creating a transformer object by PCA algorithm.}
\usage{
transformer.kpca(x, ...)
}
\arguments{
\item{components}{positive number of components for the transformed data.}

\item{scale}{boolean value to scale transformed data to the original data range.}

\item{discrete}{boolean value to handle discrete features in data.}
}
\description{
This function return a transformer object for data.
}
\examples{
data(iris)
x_trans <- transformer.kpca(iris[,1:4])

}
