% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{transformer.nmf}
\alias{transformer.nmf}
\title{Function creating a transformer object by NMF algorithm (Lee et al. 2001 based on Euclidean distance)}
\usage{
transformer.nmf(x, ...)
}
\arguments{
\item{components}{positive number of components for the transformed data, default is 2.}

\item{discrete}{boolean value to handle discrete features in data.}

\item{max_iter}{positive number to limit the iteration, default is 1000.}

\item{eps}{positive small float number to define the residual.}
}
\description{
This function return a transformer object for data.
}
\examples{
data(iris)
x_trans <- transformer.nmf(iris[,1:4])

}
