% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{transformer.pca}
\alias{transformer.pca}
\title{Function creating a transformer object using PCA algorithm.}
\usage{
transformer.pca(
  x,
  components = 2,
  center = FALSE,
  scaling = FALSE,
  handle_discrete = FALSE
)
}
\arguments{
\item{components}{positive number of components for the transformed data.}

\item{center}{boolean value to scale data. Parameter is passed to base::scale.}

\item{scaling}{boolean value to scale data. Parameter is passed to base::scale.}

\item{handle_discrete}{boolean value to handle discrete features in data.}
}
\value{
transformer.pca return a class "transformer" containing the following componenets:
}
\description{
This function return a transformer object fitted by data.
}
\details{
This calculation is created based on stat::prcomp function with some adjustments to fit into the purpose of the package. It creates a transformer object including several attribute to perform other functionality.
}
\examples{
data(iris)
x_trans <- transformer.pca(iris[,1:4])

}
